requirements.txt
streamlit
crewai
crewai-tools
langchain
langchain-community
langchain-core
langchain-openai
langchain-groq
python-dotenv